refactor(core): modularize agent into utility packages and add logging
- Move Gemini loop and tool-call JSON schema to  (, strict JSON, persisted current_context)
- Add tool server client and CLI parsing in  (GetAvailableTools, ExecuteTool, ParseAndExecuteTool)
- Introduce Slack helpers in  and Slack webhook handler in
- Add logging package  with Setup() to tee logs to stdout and debug.log; wire into startup
- Simplify : focus on HTTP API, Slack webhook wiring, DB init/migrations, graceful shutdown
- Wire dynamic tool server startup and address resolution; read API_ADDR from config
- Persist conversation context and messages via DB functions in  package
- Keep Docker/chroot constraints and JSON-only tool protocol in prompts

docs(plan): update .windsurf_plan.md with summary and next steps

Build: go build ./... passes
