+++ .windsurf_plan.md
@@ -268,3 +268,97 @@ Next:
   - Down: `migrations/0004_messages_thread_role_id_idx.down.sql`
     - `DROP INDEX IF EXISTS idx_messages_thread_role_id_desc;`
 - Rationale: speeds up `getLastContextForThread()` in `agent.go` which filters by `thread_id` and `role` and orders by `id DESC LIMIT 1`.
+
+## Small Refactor (2025-08-10)
+- Extracted path normalization helpers from `agent.go` into `path_normalize.go`:
+  - Moved: `doubleSlashRegex`, `normalizePathInText()`, `sanitizeContextFacts()`.
+  - Kept `getChrootDir()` in `agent.go` (shared cache), used by `path_normalize.go`.
+- Cleaned up imports in `agent.go` (retained `regexp` as it is used by `/tool` parsing).
+- Benefit: smaller `agent.go`, clearer separation of concerns.
+
+## Context Utils Refactor (2025-08-10)
+- Moved context-related helpers from `agent.go` to a new `utility_context.go` for maintainability:
+  - `getLastContextForThread(threadID int64) ([]string, error)`
+  - `getContextMaxItems() int` (with `sync.Once` cache vars)
+  - `mergeStringSets(base, extra []string) []string`
+- Cleaned imports and removed duplicate cache vars from `agent.go`.
+- Verified with `go build ./...`.
+
+## Gemini Refactor (2025-08-10)
+- Extracted Gemini-specific logic from `agent.go` to `gemini.go`:
+  - `callGeminiAPI(...)`
+  - `geminiAPIHandler(...)`
+  - `ToolCall` struct and `GetMergedContext()` method
+- Cleaned up imports and removed duplicates from `agent.go`.
+- Build verified: `go build ./...`.
+
+### Follow-up (2025-08-10)
+- Moved Slack helpers/handlers under `utility/` to reduce top-level clutter and decouple from `main` internals:
+  - `utility/slack.go`: `SendSlackResponse()`
+  - `utility/slack_handlers.go`: `HandleSlackMessage(...)` with injected deps (`getOrCreateAnyThread`, `geminiAPIHandler`).
+- Updated `agent.go` webhook route to call `utility.HandleSlackMessage(...)`.
+- Replaced top-level `slack.go` with a minimal stub comment (no code).
+- Build verified: `go build ./...`.
+
+## Misc & DB Utilities Refactor (2025-08-10)
+- Extracted generic helpers to `utility_misc.go`:
+  - `summarizeToolResponse()`
+  - `maskToken()`
+- Extracted DB helpers to `utility_db.go`:
+  - `createNewThread(title string) (int64, error)`
+  - `storeMessage(threadID int64, role, content string, metadata map[string]interface{}) error`
+- Removed moved functions from `agent.go` and cleaned imports.
+- Build verified: `go build ./...`.
+
+## Config & Logging Utilities Refactor (2025-08-10)
+- Extracted configuration helpers to `utility_config.go`:
+  - `loadConfig()`
+  - `getChrootDir()`
+  - Moved related cached variables (`configOnce`, `configData`, `configErr`, `chrootDirOnce`, `chrootDirCached`) and `dockerAutoRemove` handling.
+- Extracted logging setup to `internal/logging/logging.go`:
+  - `logging.Setup()` writes to stdout and `debug.log`, routes Gin logs as well.
+  - Removed old `setupLogging()` from `utility_logging.go` (now a stub pointing to internal).
+- Removed moved functions/vars from `agent.go` and cleaned imports.
+- Build verified: `go build ./...`.
+
+### Follow-up (2025-08-10)
+- Removed obsolete `utility_logging.go` after migrating to `internal/logging/logging.go` to avoid duplication.
+- Build verified: `go build ./...`.
+
+### Follow-up (2025-08-10)
+- Moved misc helpers to `utility/misc.go` as package `utility` with exported functions:
+  - `SummarizeToolResponse(success bool, data interface{}, errMsg string) string`
+  - `MaskToken(s string) string`
+- Updated `gemini.go` to import `go-thing/utility` and use `utility.SummarizeToolResponse`.
+- Moved DB helpers to `utility/db.go` as package `utility` with exported functions:
+  - `CreateNewThread(title string) (int64, error)`
+  - `StoreMessage(threadID int64, role, content string, metadata map[string]interface{}) error`
+- Updated call sites in `agent.go` and `slack.go` to use `utility` package.
+- Removed `utility_misc.go` and `utility_db.go` to avoid duplication.
+- Build verified: `go build ./...`.
+
+## Slack Utilities Refactor (2025-08-10)
+- Moved Slack-specific logic from `agent.go` to new `slack.go`:
+  - `SlackCache` struct and its `Get`/`Put` methods
+  - Per-channel lock registry: `getSlackChannelLock`, `removeSlackChannelLock`
+  - `ensureSlackThread(channel string)`
+  - `handleSlackMessage(...)`
+  - `sendSlackResponse(channel, message string)`
+- Removed the above from `agent.go` to reduce file size and improve maintainability.
+- Build verified: `go build ./...`.
+
+## Refactor (2025-08-10)
+- Moved path normalization helpers into `utility/path_normalize.go` as exported `NormalizePathInText` and `SanitizeContextFacts`.
+- Updated `gemini.go` to call `utility.SanitizeContextFacts(...)`.
+- Replaced `path_normalize.go` at repo root with a stub comment to avoid duplicate symbols.
+- Build verified: `go build ./...`.
+
+### Follow-up (2025-08-10)
+- Removed the legacy `path_normalize.go` file from the repo root after verifying all references point to `utility/` and build passes.
+
+## Gemini Refactor (2025-08-10)
+- Moved Gemini logic to `utility/gemini.go` with exported `GeminiAPIHandler(...)`.
+- Moved context limit helper to `utility/context_limits.go`.
+- Updated `agent.go` to call `utility.GeminiAPIHandler` and pass it into Slack handler.
+- Removed old root files: `gemini.go`, `context_limits.go`.
+- Build verified: `go build ./...`.
